requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
aiosqlite==0.21.0
asyncpg==0.30.0
alembic==1.17.0
pydantic==2.12.3
python-dotenv==1.2.1
pytest==8.4.2
httpx==0.28.1
